Another major refactoring introducing DatasetHelper class
