+++ src/hdx/freshness/emailer/datafreshnessstatus.py
@@ -10,149 +10,22 @@ Determines freshness status
 import datetime
 import logging
 
-from hdx.data.dataset import Dataset
-from hdx.data.organization import Organization
-from hdx.data.user import User
-from hdx.hdx_configuration import Configuration
 from hdx.utilities.dictandlist import dict_of_lists_add
 
 from hdx.freshness.emailer.freshnessemail import Email
-from hdx.freshness.emailer.utilities import get_dataset_dates
 
 logger = logging.getLogger(__name__)
 
 
 class DataFreshnessStatus:
-    freshness_status = {0: 'Fresh', 1: 'Due', 2: 'Overdue', 3: 'Delinquent'}
     object_output_limit = 2
 
-    def __init__(self, site_url, databasequeries, email, sheet, users=None, organizations=None,
-                 ignore_sysadmin_emails=None):
+    def __init__(self, datasethelper, databasequeries, email, sheet):
         ''''''
-        self.site_url = site_url
+        self.datasethelper = datasethelper
         self.databasequeries = databasequeries
-        if users is None:  # pragma: no cover
-            users = User.get_all_users()
         self.email = email
         self.sheet = sheet
-        if organizations is None:  # pragma: no cover
-            organizations = Organization.get_all_organization_names(all_fields=True, include_users=True)
-        if ignore_sysadmin_emails is None:  # pragma: no cover
-            ignore_sysadmin_emails = Configuration.read()['ignore_sysadmin_emails']
-        self.users = dict()
-        self.sysadmins = dict()
-        self.sysadmins_to_email = list()
-        for user in users:
-            userid = user['id']
-            self.users[userid] = user
-            if user['sysadmin']:
-                self.sysadmins[userid] = user
-                if user['fullname'] and user['email'] not in ignore_sysadmin_emails:
-                    self.sysadmins_to_email.append(user)
-        self.organizations = dict()
-        for organization in organizations:
-            users_per_capacity = dict()
-            for user in organization['users']:
-                dict_of_lists_add(users_per_capacity, user['capacity'], user['id'])
-            self.organizations[organization['id']] = users_per_capacity
-
-    def get_maintainer(self, dataset):
-        maintainer = dataset['maintainer']
-        return self.users.get(maintainer)
-
-    def get_org_admins(self, dataset):
-        organization_id = dataset['organization_id']
-        orgadmins = list()
-        organization = self.organizations[organization_id]
-        if 'admin' in organization:
-            for userid in self.organizations[organization_id]['admin']:
-                user = self.users.get(userid)
-                if user:
-                    orgadmins.append(user)
-        return orgadmins
-
-    def get_maintainer_orgadmins(self, dataset):
-        users_to_email = list()
-        maintainer = self.get_maintainer(dataset)
-        if maintainer is not None:
-            users_to_email.append(maintainer)
-            maintainer_name = self.get_user_name(maintainer)
-            maintainer = (maintainer_name, maintainer['email'])
-        orgadmins = list()
-        for orgadmin in self.get_org_admins(dataset):
-            if maintainer is None:
-                users_to_email.append(orgadmin)
-            username = self.get_user_name(orgadmin)
-            orgadmins.append((username, orgadmin['email']))
-        return maintainer, orgadmins, users_to_email
-
-    @staticmethod
-    def get_update_frequency(dataset):
-        if dataset['update_frequency'] is None:
-            return 'NOT SET'
-        else:
-            return Dataset.transform_update_frequency('%d' % dataset['update_frequency']).lower()
-
-    @staticmethod
-    def get_user_name(user):
-        user_name = user.get('display_name')
-        if not user_name:
-            user_name = user['fullname']
-            if not user_name:
-                user_name = user['name']
-        return user_name
-
-    def get_dataset_url(self, dataset):
-        return '%sdataset/%s' % (self.site_url, dataset['name'])
-
-    def get_organization_url(self, organization):
-        return '%sorganization/%s' % (self.site_url, organization['name'])
-
-    def create_dataset_string(self, dataset, maintainer, orgadmins, sysadmin=False, include_org=True,
-                              include_freshness=False, include_datasetdate=False):
-        url = self.get_dataset_url(dataset)
-        msg = list()
-        htmlmsg = list()
-        msg.append('%s (%s)' % (dataset['title'], url))
-        htmlmsg.append('<a href="%s">%s</a>' % (url, dataset['title']))
-        if sysadmin and include_org:
-            orgmsg = ' from %s' % dataset['organization_title']
-            msg.append(orgmsg)
-            htmlmsg.append(orgmsg)
-        if maintainer is not None:
-            if sysadmin:
-                user_name, user_email = maintainer
-                msg.append(' maintained by %s (%s)' % (user_name, user_email))
-                htmlmsg.append(' maintained by <a href="mailto:%s">%s</a>' % (user_email, user_name))
-        else:
-            if sysadmin:
-                missing_maintainer = ' with missing maintainer and organization administrators '
-                msg.append(missing_maintainer)
-                htmlmsg.append(missing_maintainer)
-
-            usermsg = list()
-            userhtmlmsg = list()
-            for orgadmin in orgadmins:
-                user_name, user_email = orgadmin
-                usermsg.append('%s (%s)' % (user_name, user_email))
-                userhtmlmsg.append('<a href="mailto:%s">%s</a>' % (user_email, user_name))
-            if sysadmin:
-                msg.append(', '.join(usermsg))
-                htmlmsg.append(', '.join(userhtmlmsg))
-        update_frequency = self.get_update_frequency(dataset)
-        msg.append(' with expected update frequency: %s' % update_frequency)
-        htmlmsg.append(' with expected update frequency: %s' % update_frequency)
-        if include_freshness:
-            fresh = self.freshness_status.get(dataset['fresh'], 'None')
-            msg.append(' and freshness: %s' % fresh)
-            htmlmsg.append(' and freshness: %s' % fresh)
-        if include_datasetdate:
-            datasetdate = dataset['dataset_date']
-            msg.append(' and date of dataset: %s' % datasetdate)
-            htmlmsg.append(' and date of dataset: %s' % datasetdate)
-        Email.output_newline(msg, htmlmsg)
-
-        return ''.join(msg), ''.join(htmlmsg)
 
     def check_number_datasets(self, now, send_failures=None):
         logger.info('\n\n*** Checking number of datasets ***')
@@ -182,7 +55,7 @@ class DataFreshnessStatus:
                 title = 'WARNING: Fall in datasets on HDX today!'
                 msg = 'Dear system administrator,\n\nThere are %d (%d%%) fewer datasets today than yesterday on HDX!\n' % \
                       (diff_datasets, percentage_diff * 100)
-                send_to = self.sysadmins_to_email
+                send_to = self.datasethelper.sysadmins_to_email
                 stop = False
         else:
             logger.info('No issues with number of datasets.')
@@ -200,30 +73,26 @@ class DataFreshnessStatus:
         msg = [startmsg]
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
 
-        def output_tabs(n=1):
-            for i in range(n):
-                msg.append('  ')
-                htmlmsg.append('&nbsp&nbsp')
-
         def create_broken_dataset_string(ds, ma, oa):
             dataset_string, dataset_html_string = \
-                self.create_dataset_string(ds, ma, oa, sysadmin=True, include_org=False, include_freshness=True)
-            output_tabs(2)
+                self.datasethelper.create_dataset_string(ds, ma, oa, sysadmin=True, include_org=False,
+                                                         include_freshness=True)
+            Email.output_tabs(msg, htmlmsg, 2)
             msg.append(dataset_string)
             htmlmsg.append(dataset_html_string)
             newline = False
             for i, resource in enumerate(sorted(ds['resources'], key=lambda d: d['name'])):
                 if i == self.object_output_limit:
-                    output_tabs(3)
+                    Email.output_tabs(msg, htmlmsg, 3)
                 if i >= self.object_output_limit:
                     newline = True
-                    output_tabs(1)
+                    Email.output_tabs(msg, htmlmsg, 1)
                     msg.append('%s (%s)' % (resource['name'], resource['id']))
                     htmlmsg.append('%s (%s)' % (resource['name'], resource['id']))
                     continue
                 resource_string = 'Resource %s (%s) has error: %s!' % \
                                   (resource['name'], resource['id'], resource['error'])
-                output_tabs(4)
+                Email.output_tabs(msg, htmlmsg, 4)
                 msg.append('%s\n' % resource_string)
                 htmlmsg.append('%s<br>' % resource_string)
             if newline:
@@ -231,41 +100,31 @@ class DataFreshnessStatus:
 
         def create_cut_down_broken_dataset_string(i, ds):
             if i == self.object_output_limit:
-                output_tabs(1)
+                Email.output_tabs(msg, htmlmsg, 1)
             if i >= self.object_output_limit:
-                url = self.get_dataset_url(ds)
-                output_tabs(1)
+                url = self.datasethelper.get_dataset_url(ds)
+                Email.output_tabs(msg, htmlmsg, 1)
                 msg.append('%s (%s)' % (ds['title'], url))
                 htmlmsg.append('<a href="%s">%s</a>' % (url, ds['title']))
                 return True
             return False
 
-        def output_error(error):
-            msg.append(error)
-            htmlmsg.append('<b>%s</b>' % error)
-            Email.output_newline(msg, htmlmsg)
-
-        def output_org(title):
-            msg.append(title)
-            htmlmsg.append('<b><i>%s</i></b>' % title)
-            Email.output_newline(msg, htmlmsg)
-
         for error_type in sorted(datasets):
-            output_error(error_type)
+            Email.output_error(msg, htmlmsg, error_type)
             datasets_error = datasets[error_type]
             for org_title in sorted(datasets_error):
-                output_org(org_title)
+                Email.output_org(msg, htmlmsg, org_title)
                 org = datasets_error[org_title]
                 newline = False
                 for i, dataset_name in enumerate(sorted(org)):
                     dataset = org[dataset_name]
-                    maintainer, orgadmins, _ = self.get_maintainer_orgadmins(dataset)
+                    maintainer, orgadmins, _ = self.datasethelper.get_maintainer_orgadmins(dataset)
                     cut_down = create_cut_down_broken_dataset_string(i, dataset)
                     if cut_down:
                         newline = True
                     else:
                         create_broken_dataset_string(dataset, maintainer, orgadmins)
-                    url = self.get_dataset_url(dataset)
+                    url = self.datasethelper.get_dataset_url(dataset)
                     title = dataset['title']
                     if maintainer:
                         maintainer_name, maintainer_email = maintainer
@@ -273,9 +132,9 @@ class DataFreshnessStatus:
                         maintainer_name, maintainer_email = '', ''
                     orgadmin_names = ','.join([x[0] for x in orgadmins])
                     orgadmin_emails = ','.join([x[1] for x in orgadmins])
-                    update_freq = self.get_update_frequency(dataset)
+                    update_freq = self.datasethelper.get_update_frequency(dataset)
                     latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
-                    fresh = self.freshness_status.get(dataset['fresh'], 'None')
+                    fresh = self.datasethelper.freshness_status.get(dataset['fresh'], 'None')
                     error = list()
                     for resource in sorted(dataset['resources'], key=lambda d: d['name']):
                         error.append('%s:%s' % (resource['name'], resource['error']))
@@ -293,7 +152,7 @@ class DataFreshnessStatus:
             Email.output_newline(msg, htmlmsg)
 
         if sendto is None:
-            users_to_email = self.sysadmins_to_email
+            users_to_email = self.datasethelper.sysadmins_to_email
         else:
             users_to_email = sendto
         self.email.close_send(users_to_email, 'Broken datasets', msg, htmlmsg)
@@ -314,11 +173,12 @@ class DataFreshnessStatus:
         msg = [startmsg]
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
         for dataset in sorted(datasets, key=lambda d: (d['organization_title'], d['name'])):
-            maintainer, orgadmins, _ = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins, sysadmin=True)
+            maintainer, orgadmins, _ = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins, sysadmin=True)
             msg.append(dataset_string)
             htmlmsg.append(dataset_html_string)
-            url = self.get_dataset_url(dataset)
+            url = self.datasethelper.get_dataset_url(dataset)
             title = dataset['title']
             org_title = dataset['organization_title']
             if maintainer:
@@ -327,7 +187,7 @@ class DataFreshnessStatus:
                 maintainer_name, maintainer_email = '', ''
             orgadmin_names = ','.join([x[0] for x in orgadmins])
             orgadmin_emails = ','.join([x[1] for x in orgadmins])
-            update_freq = self.get_update_frequency(dataset)
+            update_freq = self.datasethelper.get_update_frequency(dataset)
             latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
             # URL	Title	Organisation	Maintainer	Maintainer Email	Org Admins	Org Admin Emails
             # Update Frequency	Latest of Modifieds
@@ -336,7 +196,7 @@ class DataFreshnessStatus:
                    'Org Admins': orgadmin_names, 'Org Admin Emails': orgadmin_emails,
                    'Update Frequency': update_freq, 'Latest of Modifieds': latest_of_modifieds}
             datasets_flat.append(row)
-        self.email.close_send(self.sysadmins_to_email, 'Delinquent datasets', msg, htmlmsg)
+        self.email.close_send(self.datasethelper.sysadmins_to_email, 'Delinquent datasets', msg, htmlmsg)
         return datasets_flat
 
     def process_delinquent(self):
@@ -353,8 +213,9 @@ class DataFreshnessStatus:
         starthtmlmsg = Email.html_start(Email.convert_newlines(startmsg))
         all_users_to_email = dict()
         for dataset in sorted(datasets, key=lambda d: (d['organization_title'], d['name'])):
-            maintainer, orgadmins, users_to_email = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins)
+            maintainer, orgadmins, users_to_email = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins)
             for user in users_to_email:
                 id = user['id']
                 output_list = all_users_to_email.get(id)
@@ -364,12 +225,12 @@ class DataFreshnessStatus:
                 output_list.append((dataset_string, dataset_html_string))
         emails = dict()
         for id in sorted(all_users_to_email.keys()):
-            user = self.users[id]
-            basemsg = startmsg % self.get_user_name(user)
+            user = self.datasethelper.users[id]
+            basemsg = startmsg % self.datasethelper.get_user_name(user)
             dict_of_lists_add(emails, 'plain', basemsg)
             dict_of_lists_add(emails, 'html', self.email.convert_newlines(basemsg))
             msg = [basemsg]
-            htmlmsg = [starthtmlmsg % self.get_user_name(user)]
+            htmlmsg = [starthtmlmsg % self.datasethelper.get_user_name(user)]
             for dataset_string, dataset_html_string in all_users_to_email[id]:
                 msg.append(dataset_string)
                 htmlmsg.append(dataset_html_string)
@@ -396,12 +257,12 @@ class DataFreshnessStatus:
         msg = [startmsg]
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
         for dataset in sorted(invalid_maintainers, key=lambda d: (d['organization_title'], d['name'])):
-            maintainer, orgadmins, _ = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins,
-                                                                             sysadmin=True)
+            maintainer, orgadmins, _ = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins, sysadmin=True)
             msg.append(dataset_string)
             htmlmsg.append(dataset_html_string)
-            url = self.get_dataset_url(dataset)
+            url = self.datasethelper.get_dataset_url(dataset)
             title = dataset['title']
             org_title = dataset['organization_title']
             if maintainer:
@@ -410,7 +271,7 @@ class DataFreshnessStatus:
                 maintainer_name, maintainer_email = '', ''
             orgadmin_names = ','.join([x[0] for x in orgadmins])
             orgadmin_emails = ','.join([x[1] for x in orgadmins])
-            update_freq = self.get_update_frequency(dataset)
+            update_freq = self.datasethelper.get_update_frequency(dataset)
             latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
             # URL	Title	Organisation	Maintainer	Maintainer Email	Org Admins	Org Admin Emails
             # Update Frequency	Latest of Modifieds
@@ -419,7 +280,7 @@ class DataFreshnessStatus:
                    'Org Admins': orgadmin_names, 'Org Admin Emails': orgadmin_emails,
                    'Update Frequency': update_freq, 'Latest of Modifieds': latest_of_modifieds}
             datasets_flat.append(row)
-        self.email.close_send(self.sysadmins_to_email, 'Datasets with invalid maintainer', msg, htmlmsg)
+        self.email.close_send(self.datasethelper.sysadmins_to_email, 'Datasets with invalid maintainer', msg, htmlmsg)
         return datasets_flat
 
     def send_orgadmins_email(self, invalid_orgadmins):
@@ -432,7 +293,7 @@ class DataFreshnessStatus:
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
         for key in sorted(invalid_orgadmins):
             organization = invalid_orgadmins[key]
-            url = self.get_organization_url(organization)
+            url = self.datasethelper.get_organization_url(organization)
             title = organization['title']
             error = organization['error']
             msg.append('%s (%s)' % (title, url))
@@ -442,13 +303,15 @@ class DataFreshnessStatus:
             # URL	Title	Problem
             row = {'URL': url, 'Title': title, 'Error': error}
             organizations_flat.append(row)
-        self.email.close_send(self.sysadmins_to_email, 'Organizations with invalid admins', msg, htmlmsg)
+        self.email.close_send(self.datasethelper.sysadmins_to_email, 'Organizations with invalid admins', msg, htmlmsg)
         return organizations_flat
 
     def process_maintainer_orgadmins(self):
         logger.info('\n\n*** Checking for invalid maintainers and organisation administrators ***')
         invalid_maintainers, invalid_orgadmins = \
-            self.databasequeries.get_invalid_maintainer_orgadmins(self.organizations, self.users, self.sysadmins)
+            self.databasequeries.get_invalid_maintainer_orgadmins(self.datasethelper.organizations,
+                                                                  self.datasethelper.users,
+                                                                  self.datasethelper.sysadmins)
         datasets = self.send_maintainer_email(invalid_maintainers)
         self.sheet.update('Maintainer', datasets)
         datasets = self.send_orgadmins_email(invalid_orgadmins)
@@ -464,12 +327,12 @@ class DataFreshnessStatus:
         msg = [startmsg]
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
         for dataset in datasets:
-            maintainer, orgadmins, _ = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins,
-                                                                             sysadmin=True)
+            maintainer, orgadmins, _ = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins, sysadmin=True)
             msg.append(dataset_string)
             htmlmsg.append(dataset_html_string)
-            url = self.get_dataset_url(dataset)
+            url = self.datasethelper.get_dataset_url(dataset)
             title = dataset['title']
             org_title = dataset['organization_title']
             if maintainer:
@@ -478,7 +341,7 @@ class DataFreshnessStatus:
                 maintainer_name, maintainer_email = '', ''
             orgadmin_names = ','.join([x[0] for x in orgadmins])
             orgadmin_emails = ','.join([x[1] for x in orgadmins])
-            update_freq = self.get_update_frequency(dataset)
+            update_freq = self.datasethelper.get_update_frequency(dataset)
             latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
             # URL	Title	Organisation	Maintainer	Maintainer Email	Org Admins	Org Admin Emails
             # Update Frequency	Latest of Modifieds
@@ -487,7 +350,7 @@ class DataFreshnessStatus:
                    'Org Admins': orgadmin_names, 'Org Admin Emails': orgadmin_emails,
                    'Update Frequency': update_freq, 'Latest of Modifieds': latest_of_modifieds}
             datasets_flat.append(row)
-        self.email.close_send(self.sysadmins_to_email, 'Datasets with no resources', msg, htmlmsg)
+        self.email.close_send(self.datasethelper.sysadmins_to_email, 'Datasets with no resources', msg, htmlmsg)
         return datasets_flat
 
     def process_datasets_noresources(self):
@@ -505,9 +368,10 @@ class DataFreshnessStatus:
         starthtmlmsg = Email.html_start(Email.convert_newlines(startmsg))
         all_users_to_email = dict()
         for dataset in sorted(datasets, key=lambda d: (d['organization_title'], d['name'])):
-            maintainer, orgadmins, users_to_email = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins,
-                                                                             include_datasetdate=True)
+            maintainer, orgadmins, users_to_email = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins,
+                                                                                           include_datasetdate=True)
             for user in users_to_email:
                 id = user['id']
                 output_list = all_users_to_email.get(id)
@@ -515,7 +379,7 @@ class DataFreshnessStatus:
                     output_list = list()
                     all_users_to_email[id] = output_list
                 output_list.append((dataset_string, dataset_html_string))
-            url = self.get_dataset_url(dataset)
+            url = self.datasethelper.get_dataset_url(dataset)
             title = dataset['title']
             org_title = dataset['organization_title']
             if maintainer:
@@ -524,10 +388,10 @@ class DataFreshnessStatus:
                 maintainer_name, maintainer_email = '', ''
             orgadmin_names = ','.join([x[0] for x in orgadmins])
             orgadmin_emails = ','.join([x[1] for x in orgadmins])
-            start_date, end_date = get_dataset_dates(dataset)
+            start_date, end_date = self.datasethelper.get_dataset_dates(dataset)
             start_date = start_date.isoformat()
             end_date = end_date.isoformat()
-            update_freq = self.get_update_frequency(dataset)
+            update_freq = self.datasethelper.get_update_frequency(dataset)
             latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
             # URL	Title	Organisation	Maintainer	Maintainer Email	Org Admins	Org Admin Emails
             # Update Frequency	Latest of Modifieds
@@ -538,12 +402,12 @@ class DataFreshnessStatus:
             datasets_flat.append(row)
         emails = dict()
         for id in sorted(all_users_to_email.keys()):
-            user = self.users[id]
-            basemsg = startmsg % self.get_user_name(user)
+            user = self.datasethelper.users[id]
+            basemsg = startmsg % self.datasethelper.get_user_name(user)
             dict_of_lists_add(emails, 'plain', basemsg)
             dict_of_lists_add(emails, 'html', self.email.convert_newlines(basemsg))
             msg = [basemsg]
-            htmlmsg = [starthtmlmsg % self.get_user_name(user)]
+            htmlmsg = [starthtmlmsg % self.datasethelper.get_user_name(user)]
             for dataset_string, dataset_html_string in all_users_to_email[id]:
                 msg.append(dataset_string)
                 htmlmsg.append(dataset_html_string)
@@ -572,12 +436,12 @@ class DataFreshnessStatus:
         msg = [startmsg]
         htmlmsg = [Email.html_start(Email.convert_newlines(startmsg))]
         for dataset in datasets:
-            maintainer, orgadmins, _ = self.get_maintainer_orgadmins(dataset)
-            dataset_string, dataset_html_string = self.create_dataset_string(dataset, maintainer, orgadmins,
-                                                                             sysadmin=True)
+            maintainer, orgadmins, _ = self.datasethelper.get_maintainer_orgadmins(dataset)
+            dataset_string, dataset_html_string = self.datasethelper.create_dataset_string(dataset, maintainer,
+                                                                                           orgadmins, sysadmin=True)
             msg.append(dataset_string)
             htmlmsg.append(dataset_html_string)
-            url = self.get_dataset_url(dataset)
+            url = self.datasethelper.get_dataset_url(dataset)
             title = dataset['title']
             org_title = dataset['organization_title']
             if maintainer:
@@ -586,7 +450,7 @@ class DataFreshnessStatus:
                 maintainer_name, maintainer_email = '', ''
             orgadmin_names = ','.join([x[0] for x in orgadmins])
             orgadmin_emails = ','.join([x[1] for x in orgadmins])
-            update_freq = self.get_update_frequency(dataset)
+            update_freq = self.datasethelper.get_update_frequency(dataset)
             latest_of_modifieds = dataset['latest_of_modifieds'].isoformat()
             # URL	Title	Organisation	Maintainer	Maintainer Email	Org Admins	Org Admin Emails
             # Update Frequency	Latest of Modifieds
@@ -595,7 +459,7 @@ class DataFreshnessStatus:
                    'Org Admins': orgadmin_names, 'Org Admin Emails': orgadmin_emails,
                    'Update Frequency': update_freq, 'Latest of Modifieds': latest_of_modifieds}
             datasets_flat.append(row)
-        self.email.close_send(self.sysadmins_to_email, 'Candidates for the datagrid', msg, htmlmsg)
+        self.email.close_send(self.datasethelper.sysadmins_to_email, 'Candidates for the datagrid', msg, htmlmsg)
         return datasets_flat
 
     def process_datasets_datagrid(self, sendto=None, sysadmins=None):
